Fix a bug causing strange build issues
